Version To Present To Course
Version To Present To Course

+++ src/app/user-profile-card/user-profile-card.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { User } from '../user';
 import remote from "../services/kinvey-remote-service.service.js";
+import { ToastrService, Toast } from 'ngx-toastr';
 @Component({
   selector: 'app-user-profile-card',
   templateUrl: './user-profile-card.component.html',
@@ -9,21 +10,37 @@ import remote from "../services/kinvey-remote-service.service.js";
 export class UserProfileCardComponent implements OnInit {
   @Input() user: User;
   @Output() changedUsers = new EventEmitter<any>();
-  constructor(private remote:remote) {
+  constructor(private remote:remote, private toastr: ToastrService) {
     
+  }
+  
+  
+  testFunc(){
+    if(sessionStorage.getItem("isAdmin")==="Yes"){
+      this.remote.DeleteUser(this.user._id).subscribe((data)=>{
+        this.changedUsers.emit(this.user);
+        this.toastr.info("User Deleted")
+      })
+    }
+    else{this.toastr.error("Not Admin")}
+  }
+  
+  IsDisabled(){
+    if (this.user._kmd.status){
+      return false;
+    }
+    else return true;
   }
 
   ngOnInit() {
-    console.log("lol")
   }
-
-  testFunc(){
-    console.log(this.user._id);
-    this.remote.DeleteUser(this.user._id).subscribe((data)=>{
-      console.log("it worked lol");
-      this.changedUsers.emit(this.user);
-
-    })
-  }
-
 }
+
+      
+  
+
+
+      
+
+
+
